BARK-113: Move RegistrationService into lib/bark
